Revert "Bullet Trials!"
This reverts commit eee035982e0b9bf689da20f4b055dd1fa4578913.

+++ Assets/Scripts/DestroyByTime.cs
@@ -3,7 +3,7 @@ using System.Collections;
 
 public class DestroyByTime : MonoBehaviour {
 
-	public float seconds;
+	public int seconds;
 
 	// Use this for initialization
 	void Start () {
